컴포넌트 분리
 - 숫자 버튼에 키 값을 부여하고 구현

 - onClick 실행 시 값'1' 넘기
