+++ src/index.jsx
@@ -1,73 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-
-function Counter({ count, onClick }) {
-  return (
-    <button type="button" onClick={onClick(1)}>
-      Click me!
-      (
-      { count }
-      )
-    </button>
-  );
-}
-
-function Button({ children }) {
-  return (
-    <button type="button">
-      {children}
-    </button>
-  );
-}
-
-function Buttons() {
-  function handleClickNumber() {
-  }
-  return (
-    <p>
-      {[1, 2, 3, 4, 5].map((i) => (
-        <button type="button" onClick={() => handleClickNumber(i)}>
-          {i}
-        </button>
-      ))}
-    </p>
-  );
-}
-
-function Page({ count, onClick }) {
-  return (
-    <div>
-      <p>Hello,world!!!</p>
-      <p>Hi!</p>
-      <Counter
-        count={count}
-        onClick={onClick}
-      />
-      <Buttons onCick={handleClickNumber} />
-    </div>
-  );
-}
-
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
-
-  const { count } = state;
-
-  function handleClick(increase) {
-    setState({
-      count: count + increase,
-    });
-  }
-
-  return (
-    <Page
-      count={count}
-      onClick={handleClick}
-    />
-  );
-}
+import App from './App';
 
 ReactDOM.render(
   <App />,
